fix:Refactor codebase; rename files, update import paths
Renamed several tool scripts in the src/ibx_tools/bin directory by removing 'wapi' from their names. Updated the import paths in these scripts to reflect the new location of the 'ibx_logger' module, which was moved from the 'util' to the 'logger' directory. This makes the codebase easier to navigate and maintain.