Add Apache License file and update mkdocs navigation
An Apache License file has been added to the repository to clearly define the terms and conditions of use. Additionally, several changes were made to 'mkdocs.yml' to improve the structure and organization of the documentation: the home page was renamed to 'IBX-Tools', 'user-guide/classes' was updated for accuracy and consistency, a typo in 'Grid Backup' was fixed, and the newly added 'LICENSE.md' was included in the navigation.