docs: Update user guide for establishing connection
Detailed examples and tips were added to the user guide on how to
establish a connection to a NIOS Grid. This includes using positional
and named arguments, making a global `WAPI` instance, and using a
dictionary to instantiate the `WAPI` class. Additionally, omissions and
potential pitfalls are highlighted, providing better guidance for users.