Add validation for rotated logs availability in the get-log example.
This commit introduces a function to check if the --rotated-logs option is only set when the --log-type is set to SYSLOG. This validation was necessary to avoid potential errors or mismatches when different log types are invoked with the rotated logs option.

Testing all examples except grid restore.  Ready for merge.