docs: Update user guide to reference `Gift` class instead of `WAPI`
The documentation in `user_guide/index.md` was updated to replace
references to the `WAPI` class with the new `Gift` class. Included are
updated descriptions of how this class interacts with the Infoblox Web
API and associated HTTP methods. A tip about installing the Basic API
Toolkit into a Python Virtual Environment was also added.