Update project name and increase version number
Changes reflect the new name of the project from "Basic API Toolkit" to
"`ibx-sdk`". The same change is mirrored in the project description in
the README.md file, making it more consistent. The version number has
also been incremented to "0.2.0-RC1" in the pyproject.toml file to mark
the new development stage.